backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-dotenv
pandas
tqdm
joblib
scikit-learn
passlib[bcrypt]
python-jose[cryptography]